Move client to top-level
If it's within provider then we can't import it into another module that's going to be imported by provider.
